+++ src/org/jpc/emulator/execution/opcodes/rm/sbb_Eb_Gb_mem.java
@@ -21,9 +21,9 @@ public class sbb_Eb_Gb_mem extends Executable
     public Branch execute(Processor cpu)
     {
         Reg op2 = cpu.regs[op2Index];
-        int add = (cpu.cf() ? 1: 0);
-        cpu.flagOp1 = (byte)op1.get8(cpu);
-        cpu.flagOp2 = (byte)op2.get8();
+        int add = (cpu.cf()? 1: 0);
+        cpu.flagOp1 = op1.get8(cpu);
+        cpu.flagOp2 = op2.get8();
         cpu.flagResult = (byte)(cpu.flagOp1 - (cpu.flagOp2 + add));
         op1.set8(cpu, (byte)cpu.flagResult);
         cpu.flagIns = UCodes.SBB8;
