+++ src/org/jpc/emulator/execution/opcodes/rm/sbb_Ew_Ib_mem.java
@@ -20,9 +20,9 @@ public class sbb_Ew_Ib_mem extends Executable
 
     public Branch execute(Processor cpu)
     {
-        int add = (cpu.cf() ? 1: 0);
-        cpu.flagOp1 = (short)op1.get16(cpu);
-        cpu.flagOp2 = (short)immb;
+        int add = (cpu.cf()? 1: 0);
+        cpu.flagOp1 = op1.get16(cpu);
+        cpu.flagOp2 = immb;
         cpu.flagResult = (short)(cpu.flagOp1 - (cpu.flagOp2 + add));
         op1.set16(cpu, (short)cpu.flagResult);
         cpu.flagIns = UCodes.SBB16;
