+++ src/org/jpc/emulator/execution/opcodes/rm/sbb_Gb_Eb.java
@@ -22,9 +22,9 @@ public class sbb_Gb_Eb extends Executable
     {
         Reg op1 = cpu.regs[op1Index];
         Reg op2 = cpu.regs[op2Index];
-        int add = (cpu.cf() ? 1: 0);
-        cpu.flagOp1 = (byte)op1.get8();
-        cpu.flagOp2 = (byte)op2.get8();
+        int add = (cpu.cf()? 1: 0);
+        cpu.flagOp1 = op1.get8();
+        cpu.flagOp2 = op2.get8();
         cpu.flagResult = (byte)(cpu.flagOp1 - (cpu.flagOp2 + add));
         op1.set8((byte)cpu.flagResult);
         cpu.flagIns = UCodes.SBB8;
